refactor(chat): extract ChatAttachmentsDelegate from ChatViewModel
Move addAttachment/removeAttachment/clearAttachments into a dedicated
ChatAttachmentsDelegate that owns pendingAttachments state transitions against
the shared _uiState flow. ChatViewModel keeps 3 one-line delegating wrappers -
public API unchanged. Read sites (sendMessage) already read pendingAttachments
straight off uiState, so no delegate read API needed.

Second separable slice of the ChatViewModel god-object (after #486 search).

Verified: ktlintCheck PASS, compileDebugKotlin PASS (local SDK).
